+++ modules/template/elements/ContainerElementVariable.php
@@ -9,30 +9,39 @@
 namespace humhub\modules\custom_pages\modules\template\elements;
 
 use humhub\helpers\Html;
+use humhub\modules\custom_pages\lib\templates\twig\TwigEngine;
 use humhub\modules\custom_pages\modules\template\services\TemplateInstanceRendererService;
 use Yii;
 
+/**
+ * @property-read string $editWrapperAttributes
+ */
 class ContainerElementVariable extends BaseElementVariable
 {
+    /**
+     * @var ContainerItem[]|null Cached items
+     */
+    private ?array $_items = null;
+
+    /**
+     * @var bool Set true to don't render the container wrapper twice on edit mode when attributes are
+     *           already rendered in the wrapper tag like `<div {{ container.editWrapperAttributes }}>`
+     */
+    private bool $isEditWrapperRendered = false;
+
+    public function __construct(BaseElementContent $elementContent)
+    {
+        parent::__construct($elementContent);
+        TwigEngine::registerSandboxExtensionAllowedFunctions(static::class, [
+            'getEditWrapperAttributes',
+        ]);
+    }
+
     public function __toString()
     {
-        // Note that the editMode can be set to $this->options in this case
-        $options = [];
-
-        if (TemplateInstanceRendererService::inEditMode()) {
-            $options = array_merge([
-                'element_id' => $this->elementContent->element_id,
-                'element_content_id' => $this->elementContent->id,
-                'element_name' => $this->elementContent->element->name,
-                'element_title' => $this->elementContent->element->getTitle(),
-                'empty' => $this->elementContent->isEmpty(),
-                'default' => $this->elementContent->isDefault(),
-            ], $options);
-        }
-
         try {
             if (!$this->elementContent->isEmpty()) {
-                return $this->render2($options);
+                return $this->render();
             }
         } catch (\Exception $e) {
             return strval($e);
@@ -41,20 +50,19 @@ class ContainerElementVariable extends BaseElementVariable
         return '';
     }
 
-    private function render2($options = [])
+    private function getItems(): array
     {
-        $items = $this->elementContent->items;
-
-        if (empty($items)) {
-            if (TemplateInstanceRendererService::inEditMode()) {
-                $content = Html::tag('div', Yii::t('CustomPagesModule.model', 'Empty <br />Container'));
-                return $this->renderEditBlock($content, ['class' => 'cp-editor-container-empty']);
-            }
-            return '';
+        if ($this->_items === null) {
+            $this->_items = $this->elementContent->items;
         }
 
+        return $this->_items;
+    }
+
+    private function render(): string
+    {
         $result = '';
-        foreach ($items as $containerItem) {
+        foreach ($this->getItems() as $containerItem) {
             $result .= $containerItem->render();
         }
 
@@ -71,16 +79,44 @@ class ContainerElementVariable extends BaseElementVariable
      * @param string $content
      * @return string
      */
-    protected function renderEditBlock(string $content, array $options = []): string
+    protected function renderEditBlock(string $content): string
     {
-        if (preg_match('#<(tr).+?</\1>#is', $content)) {
-            $tagName = 'tbody';
-        } else {
-            $tagName = 'div';
+        if ($this->getItems() === []) {
+            $content = Html::tag('div', Yii::t('CustomPagesModule.model', 'Empty <br />Container'));
         }
 
-        return Html::tag($tagName, $content, array_merge([
-            'data-editor-container-id' => $this->elementContent->id,
-        ], $options));
+        if ($this->isEditWrapperRendered) {
+            return $content;
+        }
+
+        $tagName = preg_match('#<(tr).+?</\1>#is', $content) ? 'tbody' : 'div';
+
+        return Html::tag($tagName, $content, $this->getEditWrapperAttributesArray());
+    }
+
+    protected function getEditWrapperAttributesArray(): array
+    {
+        $attributes = ['data-editor-container-id' => $this->elementContent->id];
+
+        if ($this->getItems() === []) {
+            $attributes['data-editor-container-empty'] = true;
+        }
+
+        return $attributes;
+    }
+
+    /**
+     * It is used to render Twig property {{ container.editWrapperAttributes }}
+     *
+     * @return string
+     */
+    public function getEditWrapperAttributes(): string
+    {
+        if (TemplateInstanceRendererService::inEditMode()) {
+            $this->isEditWrapperRendered = true;
+            return Html::renderTagAttributes($this->getEditWrapperAttributesArray());
+        }
+
+        return '';
     }
 }