+++ modules/template/widgets/views/editItemModal.php
@@ -69,7 +69,7 @@ use humhub\widgets\modal\ModalButton;
 
     <?php if (empty($model->contentMap)) : ?>
         <div class="text-center">
-            <?= Yii::t('CustomPagesModule.view', 'This template does not contain any elements yet.') ?>
+            <?= Yii::t('CustomPagesModule.view', 'This section has no editable elements.') ?>
         </div>
     <?php endif; ?>