+++ modules/template/widgets/views/editMultipleElements.php
@@ -22,7 +22,7 @@ use humhub\widgets\modal\ModalButton;
     'footer' => empty($model->contentMap)
         ? ModalButton::cancel(Yii::t('CustomPagesModule.base', 'Back'))
         : ModalButton::cancel() .
-          ModalButton::save()
+        ModalButton::save()
             ->submit()
             ->action('editMultipleElementsSubmit', null, '#templatePageRoot'),
     'form' => ['enableClientValidation' => false],
@@ -66,7 +66,7 @@ use humhub\widgets\modal\ModalButton;
 
     <?php if (empty($model->contentMap)) : ?>
         <div class="text-center">
-            <?= Yii::t('CustomPagesModule.view', 'This template does not contain any elements yet.') ?>
+            <?= Yii::t('CustomPagesModule.view', 'This section has no editable elements.') ?>
         </div>
     <?php endif; ?>
 </div>