+++ docs/CHANGELOG.md
@@ -5,11 +5,13 @@ Changelog
 -------------------------
 - Enh #: Migration to Bootstrap 5 for HumHub 1.18
 
-1.11.0-beta.8 (Unreleased)
+1.11.0-beta.8 (July 22, 2025)
 -----------------------------
 - Enh #432: Add Yii formatter for Twig filters
 - Enh #437: Add context menu to Structure View
 - Enh #440: Add template section "CSS & JS"
+- Enh #441: Allow to render container edit wrapper attributes separately
+- Enh #442: Improve template structure view
 
 1.11.0-beta.7 (July 4, 2025)
 -----------------------------